Fixing order of service registry and device creation
Somehow, the device needs to be created after the services have been added
to the service registry. Have to investigate further what's going on when
no service has been registered. This error is independent of the service
registry member in the device being a reference or copy.

The process quits with message "terminate called without an active exception"
in constructor of DataProcessingDevice. DataSourceDevice is not effected, so
it's possibly related to a specific service not being available.